Update router.php
Bug fix. If the index route was not defined the route took the next available route. 
With this fix, if the "index" is not defined, then it will go to the "page not found" if available.
If "index" and "page not found" have not been assigned, then a blank page will be shown